Fix/pnm 01 (#82)
* fix: test case errors, added failing testOverflow4

* test: added failing cases

* fix: multiplication overflow in _remainingAllowance()

* test: added gas snapshot

+++ src/Funnel.sol
@@ -441,7 +441,7 @@ contract Funnel is IFunnel, NativeMetaTransaction, MetaTxContext, Initializable,
     function _remainingAllowance(address _owner, address _spender) private view returns (uint256) {
         RenewableAllowance memory a = rAllowance[_owner][_spender];
 
-        uint256 recovered = a.recoveryRate * uint64(block.timestamp - a.lastUpdated);
+        uint256 recovered = uint256(a.recoveryRate) * uint64(block.timestamp - a.lastUpdated);
         uint256 remainingAllowance = MathUtil.saturatingAdd(a.remaining, recovered);
 
         return remainingAllowance > a.maxAmount ? a.maxAmount : remainingAllowance;
